# Use ArrayPool for PathHelper