… src/mscorlib/corefx/System/IO/PathInternal.CaseSensitivity.cs
… src/mscorlib/corefx/System/Runtime/InteropServices/NativeBuffer.cs
… src/mscorlib/corefx/System/Runtime/InteropServices/SafeHeapHandle.cs
… src/mscorlib/corefx/System/Runtime/InteropServices/SafeHeapHandleCache.cs
… src/mscorlib/src/System/Threading/Mutex.cs