Replace special unmanaged memory pool used by PathHelper by shared ArrayPool. Custom memory pools are not good for global system performance, and the particular implementation of the unmanaged memory pool was also problematic for CoreRT port because of its dependencies.