@benaadams 2016-12-24
@dotnet-bot test Ubuntu x64 Checked Build and Test