@JeremyKuhne 2017-01-03
@jkotas  Did you get any perf numbers for this change? Reviewing it now.