@JeremyKuhne 2017-01-04
Thanks for catching the potential realloc.

From: Jan Kotas [mailto:notifications@github.com]
Sent: Tuesday, January 3, 2017 7:46 PM
To: dotnet/coreclr <coreclr@noreply.github.com>
Cc: Jeremy Kuhne <Jeremy.Kuhne@microsoft.com>; Mention <mention@noreply.github.com>
Subject: Re: [dotnet/coreclr] Use ArrayPool for PathHelper (#8719)


@jkotas commented on this pull request.

________________________________

In src/mscorlib/corefx/System/IO/PathHelper.Windows.cs<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fdotnet%2Fcoreclr%2Fpull%2F8719&data=02%7C01%7Cjeremy.kuhne%40microsoft.com%7C2694ceb8e0934b0bcba008d4345432b6%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636190983605583332&sdata=Mk8QqetdDX2B%2BMVqgdfKMzl8%2B7MWoEl6Rj%2BMakjMxxE%3D&reserved=0>:

>                  : PathInternal.DevicePrefixLength;



-            buffer = new StringBuffer(length);

+            buffer.EnsureCapacity(length + 1);

The capacity includes the null terminator. +1 is for the null terminator. It should been there before as well, I think.

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fdotnet%2Fcoreclr%2Fpull%2F8719&data=02%7C01%7Cjeremy.kuhne%40microsoft.com%7C2694ceb8e0934b0bcba008d4345432b6%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636190983605583332&sdata=Mk8QqetdDX2B%2BMVqgdfKMzl8%2B7MWoEl6Rj%2BMakjMxxE%3D&reserved=0>, or mute the thread<https://na01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FAHzkbP8FfXUXH2rrwXzpqUW8I8TL47-_ks5rOxX2gaJpZM4LVK4a&data=02%7C01%7Cjeremy.kuhne%40microsoft.com%7C2694ceb8e0934b0bcba008d4345432b6%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C636190983605583332&sdata=wMesdQuidaur6UNCwT7T1S6ibRl7pmaiwHYFKwFj72M%3D&reserved=0>.