@jkotas 2017-01-04
> perf impact 

Here is a small test that I have written: https://gist.github.com/jkotas/132e2b8fbe29b69622b322b806635504

The deltas for switching from ArrayPool alone were within noise. However, I have noticed some inefficiencies when I took a look at perfview profile to verify that I am measuring the right thing. With fixes for these inefficiencies, there are 10% - 30% improvements now:

Baseline

Small absolute 4125ms
Small relative 9953ms
Medium absolute 1781ms
Medium relative 1781ms
Long absolute 1641ms
Long relative 1719ms

With perf fixes:

Small absolute 2844ms
Small relative 7500ms
Medium absolute 1406ms
Medium relative 1609ms
Long absolute 1391ms
Long relative 1437ms