@JeremyKuhne 2017-01-04
@jkotas Thanks for doing this and validating the numbers. The perf fixes look good.